+++ src/main/java/com/aleksander/restaurant/reservation/config/ReservationGeneratorProperties.java
@@ -7,4 +7,5 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @ConfigurationProperties(prefix = "reservation.generator")
 public class ReservationGeneratorProperties {
     private int count;
+    private int countCompleted;
 }